+++ Gemfile
@@ -1,6 +1,6 @@
 source 'https://rubygems.org'
 
-gem 'jazzy'
+gem 'jazzy', '>=0.9'
 
 gem 'xcpretty'
 gem 'xcpretty-json-formatter'
@@ -8,4 +8,16 @@ gem 'xcpretty-json-formatter'
 gem 'cocoapods'
 
 gem 'danger'
+gem 'danger-auto_label'
+gem 'danger-commit_lint'
+gem 'danger-mention'
+gem 'danger-pronto'
+gem 'danger-prose'
+gem 'danger-shellcheck'
+# gem 'danger-slather'
+gem 'danger-swiftlint'
+gem 'danger-tailor'
+gem 'danger-welcome_message'
 gem 'danger-xcode_summary'
+gem 'danger-xcodebuild'
+gem 'danger-xcov'
